Move more components to plugins
